backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
langchain~=0.2.0
langchain-community~=0.2.0
langchain-groq~=0.1.0
faiss-cpu
sentence-transformers>=3.0.1
python-dotenv>=1.0.1
pypdf>=4.2.0
python-multipart>=0.0.9
langchain-huggingface>=0.0.3
